server/requirements.txt
# --- Web Framework & Server ---
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.2

# --- LangChain Ecosystem ---
langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.5

# --- AI Models & Integration ---
cohere>=5.11.0
langchain-cohere>=0.3.1
cerebras-cloud-sdk>=1.59.0

# --- Vector Database (UPDATED NAME) ---
pinecone>=6.0.0
langchain-pinecone>=0.2.0

# --- Data & PDF Processing ---
pypdf>=5.1.0
pydantic>=2.9.2
pydantic-settings>=2.6.0

# --- Databases & Cloud ---
firebase-admin>=7.1.0
motor>=3.6.0
pymongo>=4.10.1

# --- Utilities ---
python-dotenv>=1.0.1
requests>=2.32.3